v2

+++ tsh.c
@@ -55,8 +55,8 @@ typedef volatile struct Job *JobP;
 static volatile struct Job jobs[MAXJOBS];
 static int nextjid = 1; // next job ID to allocate
 
-char **dirs; //Variable to store the search paths parsed in initpath().
-int dirs_count; //Variable to store the number strings in char **dirs.
+char **dirs; // Variable to store the search paths parsed in initpath().
+int dirs_count; // Variable to store the number of strings in dirs.
 
 extern char **environ; // defined by libc
 
@@ -148,11 +148,14 @@ static void sio_reverse(char s[]);
 static size_t sio_strlen(const char s[]);
 
 /*
+ * main - Performs the read-evaluate/execute loop.
+ * 
  * Requires:
- *   <to be filled in by the student>
+ *   "argc" is an integer that indicates how many arguments were entered on the command line 
+ * 	 	when the program was started. "argv" is an array of pointers to arrays of character objects.
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Executes the read-evaluate/execute loop program.
  */
 int
 main(int argc, char **argv)
@@ -194,11 +197,11 @@ main(int argc, char **argv)
 	 * action.sa_mask TO REFLECT THE SYNCHRONIZATION REQUIRED BY YOUR
 	 * IMPLEMENTATION OF sigint_handler().
 	 */
-
 	action.sa_handler = sigint_handler;
 	action.sa_flags = SA_RESTART;
 	if (sigemptyset(&action.sa_mask) < 0)
 		unix_error("sigemptyset error");
+	// Block SIGCHLD and SIGTSTP in sigint_handler.
 	if (sigaddset(&action.sa_mask, SIGCHLD) < 0 || sigaddset(&action.sa_mask, SIGTSTP) < 0) 
 		unix_error("sigaddset error");
 	if (sigaction(SIGINT, &action, NULL) < 0)
@@ -214,7 +217,8 @@ main(int argc, char **argv)
 	action.sa_flags = SA_RESTART;
 	if (sigemptyset(&action.sa_mask) < 0)
 		unix_error("sigemptyset error");
-	if (sigaddset(&action.sa_mask, SIGCHLD) < 0 || sigaddset(&action.sa_mask, SIGINT) < 0) 
+	// Block SIGCHLD and SIGINT in sigtstp_handler.
+	if (sigaddset(&action.sa_mask, SIGCHLD) < 0 || sigaddset(&action.sa_mask, SIGINT) < 0)
 		unix_error("sigaddset error");
 	if (sigaction(SIGTSTP, &action, NULL) < 0)
 		unix_error("sigaction error");
@@ -228,6 +232,7 @@ main(int argc, char **argv)
 	action.sa_flags = SA_RESTART;
 	if (sigemptyset(&action.sa_mask) < 0)
 		unix_error("sigemptyset error");
+	// Block SIGINT and SIGTSTP in sigchld_handler.
 	if (sigaddset(&action.sa_mask, SIGINT) < 0 || sigaddset(&action.sa_mask, SIGTSTP) < 0) 
 		unix_error("sigaddset error");
 	if (sigaction(SIGCHLD, &action, NULL) < 0)
@@ -287,10 +292,12 @@ main(int argc, char **argv)
  * when we type ctrl-c (ctrl-z) at the keyboard.
  *
  * Requires:
- *   <to be filled in by the student>
+ *   "cmdline" is a NUL ('\0') terminated string with a trailing
+ *   '\n' character.  "cmdline" must contain less than MAXARGS
+ *   arguments.
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Executes the command line evaluation program.
  */
 static void
 eval(const char *cmdline)
@@ -298,57 +305,62 @@ eval(const char *cmdline)
 	pid_t pid;
 	char *argv[MAXARGS];
 	int bg = parseline(cmdline, argv);
-	//JobP job;
+	
+	// If argv is not a built-in command.
 	if (!builtin_cmd(argv)) {
-		//* sigset_t mask, prev;
-		//* sigprocmask(SIG_BLOCK, &mask, &prev); /* Block SIGCHLD */
-		if ((pid = fork()) == 0) {
-			//in child
+		sigset_t mask, prev;
+		sigemptyset(&mask);
+		sigaddset(&mask, SIGCHLD);
+		sigprocmask(SIG_BLOCK, &mask, &prev); // Block SIGCHLD.
+		// Fork a child process.
+		pid = fork();
+		if (pid < 0) {
+			unix_error("fork unsuccessful");
+		}		
+		if (pid == 0) {
+			// In the child process.
 			
-			//* sigprocmask(SIG_SETMASK, &prev, NULL); /* Unblock SIGCHLD */
+			// Reset handling of SIGINT and SIGTSTP to defaults.
+			signal(SIGINT, SIG_DFL);
+			signal(SIGTSTP, SIG_DFL);
+			sigprocmask(SIG_SETMASK, &prev, NULL); // Unblock SIGCHLD.
 			setpgid(0,0);
 
-			// char *final_path;
-			// if (strchr(cmdline, '/') != NULL || getenv("PATH") == NULL) {
-			// 	//path name
-			// 	final_path = cmdline;
-			// } else {
-			// 	//executable name
-			// 	for(int i = 0; i < dirs_count; i++) {
-			// 		char *path = dirs[i];
-			// 		char *exec = malloc(strlen(path) + strlen(argv[0]) + 1);
-			// 		strcat(exec, path);
-			// 		strcat(exec, argv[0]);
-			// 		if (access(exec, F_OK) == 0) {
-			// 			final_path = exec;
-			// 			break;
-			// 		}
-			// 	}
-			// }			
-			// int exec_status = execve(final_path, argv, environ);
-			int exec_status = execvp(argv[0], argv); //CANT USE EXECVP!
+			char *final_path = "";
+			if (strchr(argv[0], '/') != NULL || getenv("PATH") == NULL) {
+				// argv[0] is a path name.
+				final_path = strdup(argv[0]);
+			} else {
+				// argv[0] is an executable name.
+				for (int i = 0; i < dirs_count; i++) {
+					char *path = dirs[i];
+					char *exec = malloc(strlen(path) + strlen(argv[0]) + 1); // Remember to free this malloc string.
+					strcpy(exec, path);
+					strcat(exec, argv[0]);
+					if (access(exec, F_OK) == 0) {
+						final_path = exec;
+						break;
+					} else {
+						free(exec);
+					}
+				}
+			}			
+			int exec_status = execve(final_path, argv, environ);
+			free(final_path);
 
-			if(exec_status < 0) {
-				printf("%s: Command not found\n", argv[0]);
+			if (exec_status < 0) {
+				printf("%s: Command not found.\n", argv[0]);
 				exit(1);
 			}			
 		}
-		//sigprocmask(SIG_BLOCK, &mask, &prev); /* Block SIGCHLD */
+		
+		setpgid(pid, pid);
 		addjob(jobs, pid, bg ? BG : FG, cmdline);
+		sigprocmask(SIG_SETMASK, &prev, NULL); // Unblock SIGCHLD
 		if (!bg) {
 			waitfg(pid);
 		} else {
-			//print a status message like: [1] (10113) ./myspin 1 &
-			//job = getjobpid(jobs, pid);
-			//addjob(jobs, pid, bg ? BG : FG, cmdline);
-			//printf("[%d] (%d) %s", job->jid, job->pid, cmdline);
-			//printf("[%d] (%d) %s", pid2jid(pid), pid, cmdline);
-			Sio_puts("[");
-			Sio_putl(pid2jid(pid));
-			Sio_puts("] (");
-			Sio_putl(pid);
-			Sio_puts(") ");
-			Sio_puts(cmdline);
+			printf("[%d] (%d) %s", pid2jid(pid), pid, cmdline);
 		}
 	}
 
@@ -430,10 +442,10 @@ parseline(const char *cmdline, char **argv)
  *  it immediately.
  *
  * Requires:
- *   <to be filled in by the student>
+ *   "argv" is an array of pointers to arrays of character objects. The first element in the array is the command to be checked.
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Return true if the first element in the array is a built-in command, and perform the necessary actions. Return false otherwise.
  *
  * Note:
  *   In the textbook, this function has the return type "int", but "bool"
@@ -458,66 +470,71 @@ builtin_cmd(char **argv)
  * do_bgfg - Execute the built-in bg and fg commands.
  *
  * Requires:
- *   <to be filled in by the student>
+ *   "argv" is an array of pointers to arrays of character objects. 
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Implements the bg and fg built-in commands.
  */
 static void
 do_bgfg(char **argv)
 {
-
-	// Prevent an "unused parameter" warning.  REMOVE THIS STATEMENT!
-	//(void)argv;
-
 	JobP job = NULL;
 	char *cmd = argv[0]; 
 	char *param = argv[1];
 	
-	if(param == NULL) {
+	// Print out error when param is null.
+	if (param == NULL) {
 		printf("%s command requires PID or %%jobid argument\n", cmd);
 		return;
 	}
 
-	if(param[0] != '%' && !isdigit(param[0])) { //IMPROVE?
+	char *pidptr;
+	pid_t pid = strtol(param, &pidptr, 10);
+	bool pid_fail = isdigit(param[0]) && (*pidptr != '\0');
+
+	char *jidptr;
+	char *param_jid = &param[1];
+	int jid = strtol(param_jid, &jidptr, 10);
+	bool jid_fail = (param[0] == '%') && (*jidptr != '\0');
+
+	// Print out error when param is invalid.
+	if ((param[0] != '%' && !isdigit(param[0])) || pid_fail || jid_fail) {
 		printf("%s: argument must be a PID or %%jobid\n", cmd);
 		return;
 	}
 
-	if(isdigit(param[0])) {
-		pid_t pid = atoi(param);
+	// Print out error when process does not exist.
+	if (isdigit(param[0]) && (*pidptr == '\0')) {
 		job = getjobpid(jobs, pid);
-		if(job == NULL) {
+		if (job == NULL) {
 			printf("(%d): No such process\n", pid);
 			return;
 		}
 	}
-
-	if(param[0] == '%') {
-		char *param_jid = &param[1];
-		int jid = atoi(param_jid);
+	
+	// Print out error when job does not exist.
+	if ((param[0] == '%') && (*jidptr == '\0')) {
 		job = getjobjid(jobs, jid);
-		if(job == NULL) {
+		if (job == NULL) {
 			printf("%%%d: No such job\n", jid);
 			return;
 		}
 	}
 	
-	if(strcmp(cmd, "bg") == 0) {
+	// Execute the built-in bg command.
+	if (strcmp(cmd, "bg") == 0) {
 		kill(-(job->pid), SIGCONT);
 		job->state = BG;
-		Sio_puts("[");
-		Sio_putl(job->jid);
-		Sio_puts("] (");
-		Sio_putl(job->pid);
-		Sio_puts(") ");
-		Sio_puts((const char*)job->cmdline);
+		printf("[%d] (%d) %s", job->jid, job->pid, job->cmdline);
 	}
-	if(strcmp(cmd, "fg") == 0) {
+	
+	// Execute the built-in fg command.
+	if (strcmp(cmd, "fg") == 0) {
 		kill(-(job->pid), SIGCONT);
 		job->state = FG;
 		waitfg(job->pid);
 	}
+
 	return;
 }
 
@@ -525,26 +542,29 @@ do_bgfg(char **argv)
  * waitfg - Block until process pid is no longer the foreground process.
  *
  * Requires:
- *   <to be filled in by the student>
+ *   "pid" is an integer that indicates the process pid of the job.
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Waits for a foreground job to complete.
  */
 static void
 waitfg(pid_t pid)
 {
 	JobP job = getjobpid(jobs, pid);
 
-	if(!job) {
+	if (!job) {
 		return;
 	}	
 
-	//sigprocmask here: block
+	sigset_t mask, prev;
+	sigemptyset(&mask);
+	sigaddset(&mask, SIGCHLD);
+	sigprocmask(SIG_BLOCK, &mask, &prev); // Block SIGCHLD
+	// Use sigsuspend() to test whether the specified process is still running in the foreground.
 	while (job->pid == pid && job->state == FG) {
-		sleep(1);
-		//sigsuspend() ??prev
+		sigsuspend(&prev);
 	}
-	//sigprocmask here: unblock
+	sigprocmask(SIG_SETMASK, &prev, NULL); // Unblock SIGCHLD
 
 	return;
 }
@@ -557,43 +577,49 @@ waitfg(pid_t pid)
  *   "pathstr" is a valid search path.
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Parse the search path by colons into an array of directories, adding a slash to the end of each string.
  */
 static void
 initpath(const char *pathstr)
 {
-
-	// Prevent an "unused parameter" warning.  REMOVE THIS STATEMENT!
-	//(void)pathstr;
-
-	//just parse PATH by : into array of dirs bc not given executable yet
-
-	//Count number of search paths by counting number of colons + 1.
-	int count = 0;
-	for(int i = 0; i < (int)strlen(pathstr); i++) {
+	//Count number of directories by counting number of colons in the search path + 1.
+	int count = 1;		//+1 for first directory without a colon
+	for (int i = 0; i < (int)strlen(pathstr); i++) {
 		if (pathstr[i] == ':') {
 			count++;
 		}
 	}
-	count++; //+1 for final path without a colon
 
 	dirs_count = count;
 	dirs = (char **)malloc(count * sizeof(char *));
+	if (dirs == NULL) {
+		unix_error("unsuccessful malloc for dirs");
+	}
 	int size = 0;
 
-	char *path_copy = strdup(pathstr); //has malloc so need free!
-	//char *path_start = strdup(pathstr);
+	char *path_copy = strdup(pathstr);
+	if (path_copy == NULL) {
+		unix_error("unsuccessful malloc for path_copy");
+	}
+
 	while (path_copy) {
 		char *new_ptr = strsep(&path_copy, ":");
-		if(strcmp(new_ptr, "")) {
+		if (strcmp(new_ptr, "") == 0) {
+			// If the string is empty, take the current directory.
 			char *pwd = getenv("PWD");
 			char *exec = malloc(strlen(pwd) + 2);
-			strcat(exec, pwd);
+			if (exec == NULL) {
+				unix_error("unsuccessful malloc for exec");
+			}
+			strcpy(exec, pwd);
 			strcat(exec, "/");
 			dirs[size] = exec;
 		} else {
 			char *exec = malloc(strlen(new_ptr) + 2);
-			strcat(exec, new_ptr);
+			if (exec == NULL) {
+				unix_error("unsuccessful malloc for exec");
+			}
+			strcpy(exec, new_ptr);
 			strcat(exec, "/");
 			dirs[size] = exec;
 		}
@@ -616,59 +642,58 @@ initpath(const char *pathstr)
  *  currently running children to terminate.
  *
  * Requires:
- *   <to be filled in by the student>
+ *   "signum" is an integer that indicates the signal number, which in this case should equal SIGCHLD.
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Catches SIGCHILD signals.
  */
 static void
 sigchld_handler(int signum)
 {
-	(void)signum;
-	pid_t pid;
-	int status;
-	int more_chld = 1;
+	if (signum == SIGCHLD) { // Check if signum is SIGCHLD.
+		pid_t pid;
+		int status;
+		int more_chld = 1;
 
-	while(more_chld) {
-		pid = waitpid(-1, &status, WUNTRACED|WNOHANG);
-		more_chld = pid > 0;
-		if(WIFSTOPPED(status)) {
-			JobP job = getjobpid(jobs, pid);
-			if(job) {
-				job->state = ST;
-				//fprintf(stdout, "Job [%d] (%d) terminated by signal %d\n", pid2jid(pid), pid, WSTOPSIG(status));
-				Sio_puts("Job [");
-				Sio_putl(pid2jid(pid));
-				Sio_puts("] (");
-				Sio_putl(pid);
-				Sio_puts(") stopped by signal SIG");
-				Sio_puts(signame[WSTOPSIG(status)]);
-				Sio_puts("\n");
+		while (more_chld) {
+			pid = waitpid(-1, &status, WUNTRACED|WNOHANG);
+			more_chld = pid > 0;
+			
+			// Handle stopped jobs.
+			if (WIFSTOPPED(status)) {
+				JobP job = getjobpid(jobs, pid);
+				if (job) {
+					job->state = ST;
+					Sio_puts("Job [");
+					Sio_putl(pid2jid(pid));
+					Sio_puts("] (");
+					Sio_putl(pid);
+					Sio_puts(") stopped by signal SIG");
+					Sio_puts(signame[WSTOPSIG(status)]);
+					Sio_puts("\n");
+				}
 			}
-		}
-		else if(WIFEXITED(status)) {
-			deletejob(jobs, pid);
-		}		
-		else if(WIFSIGNALED(status)) {
-			int jid = pid2jid(pid);
-			int success = deletejob(jobs, pid);
-			//printf("Job [%d] (%d) terminated by signal %d\n", jid, pid, WTERMSIG(status));
-			if(success) {
-				Sio_puts("Job [");
-				Sio_putl(jid);
-				Sio_puts("] (");
-				Sio_putl(pid);
-				Sio_puts(") terminated by signal SIG");
-				Sio_puts(signame[WTERMSIG(status)]);
-				Sio_puts("\n");
-				//exit (0);
+			// Handle exited jobs.
+			else if (WIFEXITED(status)) {
+				deletejob(jobs, pid);
+			}
+			// Handle terminated jobs.
+			else if (WIFSIGNALED(status)) {
+				int jid = pid2jid(pid);
+				int success = deletejob(jobs, pid);
+				if (success) {
+					Sio_puts("Job [");
+					Sio_putl(jid);
+					Sio_puts("] (");
+					Sio_putl(pid);
+					Sio_puts(") terminated by signal SIG");
+					Sio_puts(signame[WTERMSIG(status)]);
+					Sio_puts("\n");
+				}
 			}
 		}
 	}
 
-	// while ((pid = waitpid(-1, &status, WNOHANG)) > 0) {
-	// 	deletejob(jobs, pid);
-	// }
 	return;
 }
 
@@ -678,22 +703,19 @@ sigchld_handler(int signum)
  *  to the foreground job.
  *
  * Requires:
- *   <to be filled in by the student>
+ *   "signum" is an integer that indicates the signal number, which in this case should equal SIGINT.
  *
  * Effects:
- *   <to be filled in by the student>
+ *    Catches SIGINT (ctrl-c) signals and interrupt the job.
  */
 static void
 sigint_handler(int signum)
 {
-
-	// Prevent an "unused parameter" warning.
-	(void)signum;
-	//if(signum == 2) //=numerical value of SIGINT
-
-	pid_t pid = fgpid(jobs);
-	if(pid > 0) {
-		kill(-pid, SIGINT);
+	if (signum == SIGINT) { // Check if signum is SIGINT.
+		pid_t pid = fgpid(jobs);
+		if (pid > 0) {
+			kill(-pid, SIGINT);		//Interrupt the job.
+		}
 	}
 
 	return;
@@ -705,23 +727,21 @@ sigint_handler(int signum)
  *  foreground job by sending it a SIGTSTP.
  *
  * Requires:
- *   <to be filled in by the student>
+ *   "signum" is an integer that indicates the signal number, which in this case should equal SIGTSTP.
  *
  * Effects:
- *   <to be filled in by the student>
+ *   Catches SIGTSTP (ctrl-z) signals and stops the job.
  */
 static void
 sigtstp_handler(int signum)
 {
 
-	// Prevent an "unused parameter" warning.
-	(void)signum;
-
-	pid_t pid = fgpid(jobs);
-	if(pid > 0) {
-		kill(-pid, SIGTSTP);
+	if (signum == SIGTSTP) { // Check if signum is SIGINT.
+		pid_t pid = fgpid(jobs);
+		if (pid > 0) {
+			kill(-pid, SIGTSTP);	// Stop the job.
+		}
 	}
-
 	return;
 }
